requirements.txt
llama-index
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-vector-stores-chroma
llama-index-readers-file
streamlit
chromadb
faster-whisper
pandas
openpyxl
sqlalchemy
pillow
pypdf
torch
torchvision
torchaudio
psycopg2-binary
ollama
redis
python-docx
